RISC-V: fix direct calls for resolved methods
Normally, direct calls to other methods are not supported because
of missing support for trampolines (at OMR level). However, this
is a severe limitation because most JITs call runtime services (helpers).

To workaround this limitation, if - at isel time - a direct call calls
a resolved method whose address is already known, load that address
and use an indirect jump (`jarl`) to make the call. While this may be wastefull
if caller and callee are close enough (to fit in `jal` immediate), it
is safe(r).

This allows for some demos to be made :-)

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>