RISC-V: improve debugging experience when dealing with (virtual) registers
When debugging register allocator, it is very handy if debugger prints registers
(instance of `TR::Register`) the same way they're printed in compiler logs.
This is however difficult to achieve since (virtual) register number is stored
separately in a hashtable using register as a key.

For debuggers such as GDB to retrieve this information from hashtable would mean
to either to an inferior call (problematic) or re-implement hashtable lookup
in debugger (also problematic).

This commit solves the issue by actually storing the number as a member in
`TR::Register`. Since this may increase memory consumption a little, this is
only done if `DEBUG` is `#define`d.

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>