Avoid undefined behavior in byte and short shift simplifiers
When both operands of a shift node are constants, shift simplifier
fold it into a single constant node, pre-computing the result of the
shift.

Byte- and short- simplifiers masks high 27bits (considering only low 5 bits
as a shift amount), but this still may trigger an undefined behavior since
shift amount still may be more than 7 or 15 bits.

This commit fixes this problem by only considering low 3 (for byte-shifts) or
4 (for short-shifts) bits.

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>