Introduce new custom function to format TRIL code
This commit introduces new template function `Tril::format()` that
is to be used to format TRIL code in tests designed as drop-in
replacement for `std::snprintf()` except that:

  1. support only a (used) subset of *printf-like conversion specifiers,
  2. "modifiers" like l, ll and j are ignored and
  3. used conversion specifier and a static type of actual parameter must
     match (it is not possible to format `double` using `%X`
     conversion)

Use of a template function and static types rather than *printf-like
conversion flags allows `Tril::format()` to be used in another templates
and thus will allow to further simplify test creation.

Moreover, this makes TRIL formatting independent on `std::snprintf()`
which is known to behave differently on different systems. Therefore,
this commit introduces indirection required to fix #5183 and #5324.

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>