Improve dockerfile for riscv64 cross-compilation
This commit add recipe to create RISC-V 64 sysroot to Dockerfile so
the sysroot is created as part of docker image build.

The only drawback of this solution is that it contains hardcoded
paths to RISC-V libfakeroot and libfakechroot packages, so whenever
a new version is uploaded to Debian ports repository and old one
is removed, the Dockefile has to be updated.

Maybe there's a way around it.